fix: update code review's feedbacks
- 코드 리뷰 피드백 반영
- import` 문 활용  → 유즈케이스 내부 변환 로직의 경우 중복 네이밍 이슈로 필요
- 검증 로직 service로 책임 이전 → SRP 원칙 준수
- 필터링 중 `method` → `matchType` 으로 변수명 변경
- 유즈케이스 내부의 `@Schema` 어노테이션 제거
