+++ src/main/kotlin/com/dobby/backend/domain/model/experiment/CustomFilter.kt
@@ -6,7 +6,7 @@ import com.dobby.backend.infrastructure.database.entity.enums.areaInfo.Area
 import com.dobby.backend.infrastructure.database.entity.enums.areaInfo.Region
 
 data class CustomFilter(
-    val method: MatchType?,
+    val matchType: MatchType?,
     val studyTarget: StudyTarget?,
     val locationTarget: LocationTarget?,
     val recruitDone: Boolean?
